:hammer: refactor: Iniciando refactor, nueva arquitectura de clases
